Split TCPPing into per-command modules
